// angular/src/app/common/models/funcionalidade.models.ts
export enum Funcionalidade {
  PERMISSAO_CONSULTA = 'PERMISSAO_CONSULTA',
  PERMISSAO_CADASTRO = 'PERMISSAO_CADASTRO',
  PERMISSAO_ALTERACAO = 'PERMISSAO_ALTERACAO',
  CLIENTE_CONSULTA = 'CLIENTE_CONSULTA',
  CLIENTE_CADASTRO = 'CLIENTE_CADASTRO',
  CLIENTES_ALTERACAO = 'CLIENTES_ALTERACAO',
  COLABORADOR_CONSULTA = 'COLABORADOR_CONSULTA',
  COLABORADOR_CADASTRO = 'COLABORADOR_CADASTRO',
  COLABORADOR_ALTERACAO = 'COLABORADOR_ALTERACAO',
  COLABORADOR_ENTREVISTA_CONSULTA = 'COLABORADOR_ENTREVISTA_CONSULTA',
  COLABORADOR_ENTREVISTA_CADASTRO = 'COLABORADOR_ENTREVISTA_CADASTRO',
  COLABORADOR_ENTREVISTA_ALTERACAO = 'COLABORADOR_ENTREVISTA_ALTERACAO',
  COLABORADOR_OCORRENCIA_CADASTRO = 'COLABORADOR_OCORRENCIA_CADASTRO',
  COLABORADOR_INCOMPATIVEL_CADASTRO = 'COLABORADOR_INCOMPATIVEL_CADASTRO',
  COLABORADOR_CONTAS_CONSULTA = 'COLABORADOR_CONTAS_CONSULTA',
  COLABORADOR_CONTAS_CADASTRO = 'COLABORADOR_CONTAS_CADASTRO',
  COLABORADOR_CONTAS_ALTERACAO = 'COLABORADOR_CONTAS_ALTERACAO',
  COLABORADOR_CONTAS_EXCLUSAO = 'COLABORADOR_CONTAS_EXCLUSAO',
  COLABORADOR_MIDIAS_CONSULTA = 'COLABORADOR_MIDIAS_CONSULTA',
  COLABORADOR_MIDIAS_CADASTRO = 'COLABORADOR_MIDIAS_CADASTRO',
  COLABORADOR_MIDIAS_ALTERACAO = 'COLABORADOR_MIDIAS_ALTERACAO',
  COLABORADOR_MIDIAS_EXCLUSAO = 'COLABORADOR_MIDIAS_EXCLUSAO',
  PROSPECT_CONSULTA = 'PROSPECT_CONSULTA',
  PROSPECT_CADASTRO = 'PROSPECT_CADASTRO',
  PROSPECT_ALTERACAO = 'PROSPECT_ALTERACAO',
  EMPRESA_CONSULTA = 'EMPRESA_CONSULTA',
  EMPRESA_ALTERACAO = 'EMPRESA_ALTERACAO',
  ESCALA_CONSULTA = 'ESCALA_CONSULTA',
  ESCALA_PROGRAMACAO = 'ESCALA_PROGRAMACAO',
  USUARIOS_CONSULTA = 'USUARIOS_CONSULTA',
  USUARIOS_CADASTRO = 'USUARIOS_CADASTRO',
  USUARIOS_ALTERACAO = 'USUARIOS_ALTERACAO',
  USUARIOS_EXCLUSAO = 'USUARIOS_EXCLUSAO',
}